ADTypes.jl

Documentation for ADTypes.jl.

ADTypes.ADTypes — Module

ADTypes.jl

ADTypes.jl is a multi-valued logic system to choose an automatic differentiation (AD) package and specify its parameters.

ADTypes.AbstractADType — Type

AbstractADType

Abstract supertype for all AD choices.

Dense AD

Forward mode

Algorithmic differentiation:

ADTypes.AutoForwardDiff — Type

AutoForwardDiff{chunksize,T}

Struct used to select the ForwardDiff.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoForwardDiff(; chunksize=nothing, tag=nothing)

Type parameters

chunksize: the preferred chunk size to evaluate several derivatives at once

Fields

tag::T: a custom tag to handle nested differentiation calls (usually not necessary)

ADTypes.AutoPolyesterForwardDiff — Type

AutoPolyesterForwardDiff{chunksize,T}

Struct used to select the PolyesterForwardDiff.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoPolyesterForwardDiff(; chunksize=nothing, tag=nothing)

Type parameters

chunksize: the preferred chunk size to evaluate several derivatives at once

Fields

tag::T: a custom tag to handle nested differentiation calls (usually not necessary)

Finite differences:

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff{T1,T2,T3}

Struct used to select the FiniteDiff.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoFiniteDiff(;
    fdtype=Val(:forward), fdjtype=fdtype, fdhtype=Val(:hcentral),
    relstep=nothing, absstep=nothing, dir=true
)

Fields

fdtype::T1: finite difference type

fdjtype::T2: finite difference type for the Jacobian

fdhtype::T3: finite difference type for the Hessian

relstep: relative finite difference step size

absstep: absolute finite difference step size

dir: direction of the finite difference step

ADTypes.AutoFiniteDifferences — Type

AutoFiniteDifferences{T}

Struct used to select the FiniteDifferences.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoFiniteDifferences(; fdm)

Fields

fdm::T: a FiniteDifferenceMethod

Taylor mode:

ADTypes.AutoGTPSA — Type

AutoGTPSA{D}

Struct used to select the GTPSA.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoGTPSA(; descriptor=nothing)

Fields

descriptor::D: can be either

a GTPSA Descriptor specifying the number of variables/parameters, parameter order, individual variable/parameter truncation orders, and maximum order. See the GTPSA.jl documentation for more details.

nothing to automatically use a Descriptor given the context.

ADTypes.AutoTaylorDiff — Type

AutoTaylorDiff{order}

Struct used to select the TaylorDiff.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoTaylorDiff(; order = 1)

Type parameters

order: the order of the Taylor-mode automatic differentiation

Reverse mode

ADTypes.AutoMooncake — Type

AutoMooncake

Struct used to select the Mooncake.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoMooncake(; config=nothing)

Fields

config: either nothing or an instance of Mooncake.Config – see the docstring of Mooncake.Config for more information. AutoMooncake(; config=nothing) is equivalent to AutoMooncake(; config=Mooncake.Config()), i.e. the default configuration.

ADTypes.AutoReverseDiff — Type

AutoReverseDiff{compile}

Struct used to select the ReverseDiff.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoReverseDiff(; compile::Union{Val, Bool} = Val(false))

Fields

compile::Union{Val, Bool}: whether to allow pre-recording and reusing a tape (which speeds up the differentiation process).

If compile=false or compile=Val(false), a new tape must be recorded at every call to the differentiation operator.

If compile=true or compile=Val(true), a tape can be pre-recorded on an example input and then reused at every differentiation call.

The boolean version of this keyword argument is taken as the type parameter.

Warning

Pre-recording a tape only captures the path taken by the differentiated function when executed on the example input. If said function has value-dependent branching behavior, reusing pre-recorded tapes can lead to incorrect results. In such situations, you should keep the default setting compile=Val(false). For more details, please refer to ReverseDiff's AbstractTape API documentation.

Info

Despite what its name may suggest, the compile setting does not prescribe whether or not the tape is compiled with ReverseDiff.compile after being recorded. This is left as a private implementation detail.

ADTypes.AutoTracker — Type

AutoTracker

Struct used to select the Tracker.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoTracker()

ADTypes.AutoZygote — Type

AutoZygote

Struct used to select the Zygote.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoZygote()

Forward or reverse mode

ADTypes.AutoEnzyme — Type

AutoEnzyme{M,A}

Struct used to select the Enzyme.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoEnzyme(; mode::M=nothing, function_annotation::Type{A}=Nothing)

Type parameters

A determines how the function f to differentiate is passed to Enzyme. It can be:

a subtype of EnzymeCore.Annotation (like EnzymeCore.Const or EnzymeCore.Duplicated) to enforce a given annotation

Nothing to simply pass f and let Enzyme choose the most appropriate annotation

Fields

mode::M determines the autodiff mode (forward or reverse). It can be:

an object subtyping EnzymeCore.Mode (like EnzymeCore.Forward or EnzymeCore.Reverse) if a specific mode is required

nothing to choose the best mode automatically

ADTypes.AutoChainRules — Type

AutoChainRules{RC}

Struct used to select an automatic differentiation backend based on ChainRulesCore.jl (see the list here).

Defined by ADTypes.jl.

Constructors

AutoChainRules(; ruleconfig)

Fields

ruleconfig::RC: a ChainRulesCore.RuleConfig object.

ADTypes.AutoDiffractor — Type

AutoDiffractor

Struct used to select the Diffractor.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoDiffractor()

Symbolic mode

ADTypes.AutoFastDifferentiation — Type

AutoFastDifferentiation

Struct used to select the FastDifferentiation.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoFastDifferentiation()

ADTypes.AutoSymbolics — Type

AutoSymbolics

Struct used to select the Symbolics.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoSymbolics()

Sparse AD

ADTypes.AutoSparse — Type

AutoSparse{D,S,C}

Wraps an ADTypes.jl object to deal with sparse Jacobians and Hessians.

Fields

dense_ad::D: the underlying AD package, subtyping AbstractADType

sparsity_detector::S: the sparsity pattern detector, subtyping AbstractSparsityDetector

coloring_algorithm::C: the coloring algorithm, subtyping AbstractColoringAlgorithm

Constructors

AutoSparse(
    dense_ad;
    sparsity_detector=ADTypes.NoSparsityDetector(),
    coloring_algorithm=ADTypes.NoColoringAlgorithm()
)

ADTypes.dense_ad — Function

dense_ad(ad::AutoSparse)::AbstractADType
dense_ad(ad::AbstractADType)::AbstractADType

Return the underlying AD package for a sparse AD choice, act as the identity on a dense AD choice.

See also

AutoSparse

Sparsity detector

ADTypes.sparsity_detector — Function

sparsity_detector(ad::AutoSparse)::AbstractSparsityDetector

Return the sparsity pattern detector for a sparse AD choice.

See also

AutoSparse

AbstractSparsityDetector

ADTypes.AbstractSparsityDetector — Type

AbstractSparsityDetector

Abstract supertype for sparsity pattern detectors.

Required methods

jacobian_sparsity

hessian_sparsity

ADTypes.jacobian_sparsity — Function

jacobian_sparsity(f, x, sd::AbstractSparsityDetector)::AbstractMatrix{Bool}
jacobian_sparsity(f!, y, x, sd::AbstractSparsityDetector)::AbstractMatrix{Bool}

Use detector sd to construct a (typically sparse) matrix S describing the pattern of nonzeroes in the Jacobian of f (resp. f!) applied at x (resp. (y, x)).

ADTypes.hessian_sparsity — Function

hessian_sparsity(f, x, sd::AbstractSparsityDetector)::AbstractMatrix{Bool}

Use detector sd to construct a (typically sparse) matrix S describing the pattern of nonzeroes in the Hessian of f applied at x.

ADTypes.NoSparsityDetector — Type

NoSparsityDetector <: AbstractSparsityDetector

Trivial sparsity detector, which always returns a full sparsity pattern (only ones, no zeroes).

See also

AbstractSparsityDetector

ADTypes.KnownJacobianSparsityDetector — Type

KnownJacobianSparsityDetector(jacobian_sparsity::AbstractMatrix) <: AbstractSparsityDetector

Trivial sparsity detector used to return a known Jacobian sparsity pattern.

See also

AbstractSparsityDetector

KnownHessianSparsityDetector

ADTypes.KnownHessianSparsityDetector — Type

KnownHessianSparsityDetector(hessian_sparsity::AbstractMatrix) <: AbstractSparsityDetector

Trivial sparsity detector used to return a known Hessian sparsity pattern.

See also

AbstractSparsityDetector

KnownJacobianSparsityDetector

Coloring algorithm

ADTypes.coloring_algorithm — Function

coloring_algorithm(ad::AutoSparse)::AbstractColoringAlgorithm

Return the coloring algorithm for a sparse AD choice.

See also

AutoSparse

AbstractColoringAlgorithm

ADTypes.AbstractColoringAlgorithm — Type

AbstractColoringAlgorithm

Abstract supertype for Jacobian/Hessian coloring algorithms.

Required methods

column_coloring

row_coloring

symmetric_coloring

Note

The terminology and definitions are taken from the following paper:

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Assefaw Hadish Gebremedhin, Fredrik Manne, and Alex Pothen (2005)

ADTypes.column_coloring — Function

column_coloring(M::AbstractMatrix, ca::ColoringAlgorithm)::AbstractVector{<:Integer}

Use algorithm ca to construct a structurally orthogonal partition of the columns of M.

The result is a coloring vector c of length size(M, 2) such that for every non-zero coefficient M[i, j], column j is the only column of its color c[j] with a non-zero coefficient in row i.

ADTypes.row_coloring — Function

row_coloring(M::AbstractMatrix, ca::ColoringAlgorithm)::AbstractVector{<:Integer}

Use algorithm ca to construct a structurally orthogonal partition of the rows of M.

The result is a coloring vector c of length size(M, 1) such that for every non-zero coefficient M[i, j], row i is the only row of its color c[i] with a non-zero coefficient in column j.

ADTypes.symmetric_coloring — Function

symmetric_coloring(M::AbstractMatrix, ca::ColoringAlgorithm)::AbstractVector{<:Integer}

Use algorithm ca to construct a symmetrically structurally orthogonal partition of the columns (or rows) of the symmetric matrix M.

The result is a coloring vector c of length size(M, 1) == size(M, 2) such that for every non-zero coefficient M[i, j], at least one of the following conditions holds:

column j is the only column of its color c[j] with a non-zero coefficient in row i;

column i is the only column of its color c[i] with a non-zero coefficient in row j.

ADTypes.NoColoringAlgorithm — Type

NoColoringAlgorithm <: AbstractColoringAlgorithm

Trivial coloring algorithm, which always returns a different color for each matrix column/row.

See also

AbstractColoringAlgorithm

Modes

ADTypes.mode — Function

mode(ad::AbstractADType)

Return the differentiation mode of ad, as a subtype of AbstractMode.

ADTypes.AbstractMode — Type

AbstractMode

Abstract supertype for the traits identifying differentiation modes.

Subtypes

ForwardMode

ReverseMode

ForwardOrReverseMode

SymbolicMode

ADTypes.ForwardMode — Type

ForwardMode

Trait for AD choices that rely on forward mode algorithmic differentiation or finite differences.

These two paradigms are classified together because they can both efficiently compute Jacobian-vector products.

ADTypes.ForwardOrReverseMode — Type

ForwardOrReverseMode

Trait for AD choices that can work either in ForwardMode or ReverseMode, depending on their configuration.

Warning

This trait should rarely be used, because more precise dispatches to ForwardMode or ReverseMode should be defined.

ADTypes.ReverseMode — Type

ReverseMode

Trait for AD choices that rely on reverse mode algorithmic differentiation.

ADTypes.SymbolicMode — Type

SymbolicMode

Trait for AD choices that rely on symbolic differentiation.

Miscellaneous

ADTypes.Auto — Function

ADTypes.Auto(package::Symbol)

A shortcut that converts an AD package name into an instance of AbstractADType, with all parameters set to their default values.

Warning

This function is type-unstable by design and might lead to suboptimal performance. In most cases, you should never need it: use the individual backend types directly.

Example

import ADTypes
backend = ADTypes.Auto(:Zygote)

# output

ADTypes.AutoZygote()

Deprecated

ADTypes.AutoTapir — Type

AutoTapir

Danger

AutoTapir is deprecated following a package renaming, please use AutoMooncake instead.

Struct used to select the Tapir.jl backend for automatic differentiation.

Defined by ADTypes.jl.

Constructors

AutoTapir(; safe_mode=true)

Fields

safe_mode::Bool: whether to run additional checks to catch errors early.